+++ src/main/java/re/smartcity/energynet/component/data/client/SmallConsumerSpecification.java
@@ -20,7 +20,7 @@ public final class SmallConsumerSpecification {
 
     private double energy; // максимальная мощность в МВт
 
-    private int carbon = 0; // выброс CO2 (экология)
+    private double carbon = 0.0; // выброс CO2 (экология)
 
     private SupportedPriceCategories catprice = SupportedPriceCategories.CATEGORY_1; // ценовая категория
 
@@ -33,7 +33,7 @@ public final class SmallConsumerSpecification {
         if (data.isUseforecast() && data.getForecast() == null) {
             throw new IllegalArgumentException(Messages.ER_2);
         }
-        if (data.getCarbon() < 0) {
+        if (data.getCarbon() < 0.0) {
             throw new IllegalArgumentException(Messages.ER_4);
         }
     }
